requirements.txt
lightning[pytorch-extra]
torchvision
opencv-python
pillow
wavemix
wandb
timm